// src/app/models/User.ts
export interface User {
    roles: any;
    idUser?: number;
    firstName: string;
    lastName: string;
    email: string;
    companyName: string;
    sector: string;
    industry: string;
    revenue: number;
    headquarters: string;
    currency: string;
    profileComplete: boolean;
    baseYear: number;
    endYear: number;
    reductionAmbition: number;
  }
  